Add offers page
This commit adds a basic flocx service, REST API, and JavaScript service
client as outlined in the horizon plugin docs as well as a basic offer list
page on which to render the data. In addition, the page header was changed
from the horizon's built-in header to a custom angular component.